amc2c: cleaned up the motorhal and performed some basic tests
