+++ emBODY/eBcode/arch-arm/board/amc2c/application/src/app-board-amc2c/embot_app_board_amc2c_test.cpp
@@ -20,7 +20,10 @@
 
 #include "embot_hw_timer.h"
 
-#include "motorhal.h"
+#include "adcm.h"
+#include "enc.h"
+#include "hall.h"
+#include "pwm.h"
 
 #include "embot_hw_motor.h"
