+++ emBODY/eBcode/arch-arm/board/amc2c/bsp/motorhal/adcm.c
@@ -59,7 +59,7 @@ extern ADC_HandleTypeDef hadc2;
 // add in here minimal changes
 // --------------------------------------------------------------------------------------------------------------------
 
-void s_adcm_on_acquisition_of_currents(int16_t c1, int16_t c2, int16_t c3); 
+extern adcm_fp_int16_int16_uint16_t adcm_FP_on_acquisition_of_currents = NULL;
 
 #endif // #if defined(MOTORHAL_changes)
 
@@ -119,8 +119,11 @@ static void adcMotGetSamples(const DualModeAdcData_t *pBuf)
 /*DEBUG*/HAL_GPIO_WritePin(TP4_GPIO_Port, TP4_Pin, GPIO_PIN_RESET);
 #endif
     
-#if defined(MOTORHAL_changes)    
-    s_adcm_on_acquisition_of_currents(adcMotCurrents[0], adcMotCurrents[1], adcMotCurrents[2]);
+#if defined(MOTORHAL_changes) 
+    if(NULL != adcm_FP_on_acquisition_of_currents)
+    {        
+        adcm_FP_on_acquisition_of_currents(adcMotCurrents[0], adcMotCurrents[1], adcMotCurrents[2]);
+    }
 #endif    
 }
 
@@ -132,7 +135,7 @@ static void adcMotGetSamples(const DualModeAdcData_t *pBuf)
  * @param
  * @return
  */
-static void AdcMotHalfTransfer_cb(ADC_HandleTypeDef *hadc)
+void AdcMotHalfTransfer_cb(ADC_HandleTypeDef *hadc)
 {
     /* Record the first half of the buffer */
     adcMotGetSamples(&(AdcMotRawData[0]));
@@ -144,7 +147,7 @@ static void AdcMotHalfTransfer_cb(ADC_HandleTypeDef *hadc)
  * @param
  * @return
  */
-static void AdcMotTransferComplete_cb(ADC_HandleTypeDef *hadc)
+void AdcMotTransferComplete_cb(ADC_HandleTypeDef *hadc)
 {
     /* Record the second half of the buffer */
     adcMotGetSamples(&(AdcMotRawData[NUMBER_OF_ADC_CHANNELS]));
@@ -254,226 +257,9 @@ void AdcMotDeInit(void)
 }
 
 
-
 #if defined(MOTORHAL_changes)
 
-
-// --------------------------------------------------------------------------------------------------------------------
-// add in here all the new code required by MOTORHAL
-// --------------------------------------------------------------------------------------------------------------------
-
-#include "embot_core.h"
-#include <array>
-
-struct Converter
-{
-    static int32_t raw2current(int32_t r)
-    {
-        // return CIN_GAIN * vref_mV * r >> 16u;
-        return r;       
-    }
-    
-    static int32_t current2raw(int32_t c)
-    {
-        // int32_t tmp = (curr << 16) / (CIN_GAIN * vref_mV);
-        // return tmp;
-        return c;       
-    }    
-    
-    constexpr Converter() = default;
-};
-
-
-struct Calibrator
-{
-    static constexpr uint8_t _shift {10};
-    static constexpr size_t _maxcount {1 << _shift};
-    
-    volatile bool _done {false};
-    volatile size_t _count {0};
-    adcm_Currents _currents {};
-    std::array<int64_t, 3> cumulativerawvalues {0, 0, 0};
-    
-    constexpr Calibrator() = default;
-    
-    void init()
-    {
-       _count = 0;
-       _done = false; 
-       cumulativerawvalues.fill(0); 
-       adcm_set_ADC_callback({oncurrents, this});
-    }
-    
-    void stop()
-    {
-        adcm_set_ADC_callback({});
-        _count = 0;
-       _done = true;  
-    }
-    
-    
-    bool wait(const embot::core::relTime timeout = 100*embot::core::time1millisec)
-    {
-        bool ret {false};
-        
-        volatile embot::core::Time endtime = embot::core::now() + timeout;
-        
-        for(;;)
-        {
-            volatile embot::core::Time n = embot::core::now();
-            if(true == _done)
-            { 
-                ret = true; 
-                break; 
-            }
-            else if(n > endtime)
-            {
-                ret = false; 
-                stop();
-                break;
-            }
-        }
-
-        return ret;
-    }   
-
-
-    static void oncurrents(void *owner, const adcm_Currents * const currents)
-    {
-        Calibrator *o = reinterpret_cast<Calibrator*>(owner);   
-        
-        // i use Converter::current2raw() because technically adcm_Currents does not contain the raw ADC values but transformed values
-        o->cumulativerawvalues[0] += Converter::current2raw(currents->u);
-        o->cumulativerawvalues[1] += Converter::current2raw(currents->v);
-        o->cumulativerawvalues[2] += Converter::current2raw(currents->w);
-        
-        o->_count++;
-        
-        if(o->_count >= Calibrator::_maxcount)
-        {
-            // time to do business: prepare average currents, impose the offset to ADC, stop the calibrator  
-            
-            // dont use the >> _shift because ... 1. this operation is done only once, so who bother. 
-            //                                    2. cumulativerawvalues is int64_t. it may be (dont think so) negative so it is too complicate to optimize
-            // impose offset to adc.
-            volatile int64_t cc[3] {0, 0, 0};
-            cc[0] = o->cumulativerawvalues[0] / _maxcount;
-            cc[1] = o->cumulativerawvalues[1] / _maxcount;
-            cc[2] = o->cumulativerawvalues[2] / _maxcount;
-            AdcMotSetOffset(ADCMOT_PHASE_U, AdcMotGetOffset(ADCMOT_PHASE_U) + (o->cumulativerawvalues[0] / _maxcount));
-            AdcMotSetOffset(ADCMOT_PHASE_V, AdcMotGetOffset(ADCMOT_PHASE_V) + (o->cumulativerawvalues[1] / _maxcount));
-            AdcMotSetOffset(ADCMOT_PHASE_W, AdcMotGetOffset(ADCMOT_PHASE_W) + (o->cumulativerawvalues[2] / _maxcount));
-            
-            // stop
-            o->stop();
-            
-            return;
-        }                     
-    }    
-    
-};
-
-struct adcm_Internals
-{
-    adcm_Configuration config {};
-    Calibrator calibrator {};
-        
-    static void dummy_adc_callback(void *owner, const adcm_Currents * const currents) {}  
-    static constexpr adcm_ADC_callback dummyADCcbk { dummy_adc_callback, nullptr };         
-        
-    adcm_Internals() = default;    
-};
-
-adcm_Internals _adcm_internals {};
-    
-    
-void s_adcm_on_acquisition_of_currents(int16_t c1, int16_t c2, int16_t c3)
-{
-    // important note: {c1, c2, c3} are the raw ADC acquisitions as in adcMotCurrents[0, 1, 2]
-    // the callback requires converted currents. but to be fair in adcm.c there is no concept of how to convert them as in the amcbldc
-    // so, boh ... for now i do nothing    
-    adcm_Currents currents = {Converter::raw2current(c1), Converter::raw2current(c2), Converter::raw2current(c3)};
-    _adcm_internals.config.onadc.execute(&currents);         
-}    
-
-    
-bool adcm_Init(const adcm_Configuration &config)
-{
-    bool r {true};
-    
-    if(false == config.isvalid())
-    {
-        return false;
-    }
-    
-    _adcm_internals.config = config;
-    
-    // 1. at first we want to be sure that everything is stopped, so:
-    adcm_DeInit();  
-
-    // 2. then we init w/out calibration or w/ calibration
-    if(_adcm_internals.config.calibration == adcm_Configuration::CALIBRATION::none)        
-    {
-        HAL_StatusTypeDef hr = AdcMotInit();      
-        r = (HAL_ERROR == hr) ? false : true;       
-    }
-    else
-    {
-        // 2.b w/ calibration
-        _adcm_internals.calibrator.init();
-        // in here we need to have zero pwm, so we force it
-        PwmPhaseSet(0, 0, 0);
-        PwmPhaseEnable(PWM_PHASE_NONE);
-        
-        AdcMotInit(); // this call starts the ADC acquisition. 
-        // which in turn calls Calibrator::oncurrents() which eventally calls stop() and allows stop() to return
-        r = _adcm_internals.calibrator.wait();
-        r = true;
-    }
-    
-    return r;   
-}  
-
-bool adcm_Calibrate(uint64_t usectimeout)
-{
-    bool r {true};
-    
-        _adcm_internals.calibrator.init();
-        // in here we need to have zero pwm, so we force it
-        PwmPhaseSet(0, 0, 0);
-        PwmPhaseEnable(PWM_PHASE_NONE);
-    
-        r = _adcm_internals.calibrator.wait();
-
-    
-    return r;   
-}    
-
-
-void adcm_set_ADC_callback(const adcm_ADC_callback &cbk)
-{
-    // maybe better to protect execution of s_adcm_ADC_cbk by ...
-    // temporarily suspending the AdcMotTransferComplete_cb() and AdcMotHalfTransfer_cb()  
-    HAL_ADC_UnRegisterCallback(&hadc1, HAL_ADC_CONVERSION_COMPLETE_CB_ID);
-    HAL_ADC_UnRegisterCallback(&hadc1, HAL_ADC_CONVERSION_HALF_CB_ID);
-    
-    _adcm_internals.config.onadc = cbk.isvalid() ? cbk : _adcm_internals.dummyADCcbk;
-    
-    HAL_ADC_RegisterCallback(&hadc1, HAL_ADC_CONVERSION_COMPLETE_CB_ID, AdcMotTransferComplete_cb);
-    HAL_ADC_RegisterCallback(&hadc1, HAL_ADC_CONVERSION_HALF_CB_ID, AdcMotHalfTransfer_cb);    
-}
-
-// as AdcMotDeInit() but ... see the changes in analog_DeInit() of amcbldc
-bool adcm_DeInit()
-{
-    bool r = true;    
-    // AdcMotDeInit() unchanged is OK 
-    AdcMotDeInit();
-    // i also clear the callback
-    adcm_set_ADC_callback({});    
-    return r;    
-}
-
+// nothing else is required
 
 #endif // #if defined(MOTORHAL_changes)
