+++ emBODY/eBcode/arch-arm/board/amc2c/bsp/motorhal/pwm.c
@@ -413,84 +413,11 @@ void PwmTest(void)
 
 #endif // #if defined(MOTORHALCONFIG_DONTUSE_TESTS)
 
-
 #if defined(MOTORHAL_changes)
 
+// nothing else is required
 
-struct pwm_Internals
-{
-    volatile bool started {false};
-    volatile bool calibrating {false};
-    pwm_Configuration config {};
-        
-    void pwmcomparevalue_protect(bool on)
-    {
-        // if pwm timer is not started ... we dont protect
-        if(false == started)
-        {
-            return;
-        }
-        
-        if(on == true)
-        {
-            __HAL_TIM_DISABLE_IT(&htim1, TIM_IT_UPDATE);
-        }
-        else
-        {
-            __HAL_TIM_ENABLE_IT(&htim1, TIM_IT_UPDATE);      
-        }
-    }    
-        
-    pwm_Internals() = default;   
-};
-
-pwm_Internals _pwm_internals {};
-
-extern void pwm_Init(const pwm_Configuration &config)
-{
-    _pwm_internals.config =  config;
-    _pwm_internals.calibrating = false;
-
-    PwmInit();
-    
-    _pwm_internals.started = true;
-    
-    // meglio essere sicuri ... vedi dentro PwmInit()
-    __HAL_TIM_DISABLE_IT(&htim1, TIM_IT_BREAK);
-}
-
-extern void pwm_DeInit()
-{
-    _pwm_internals.started = false;
-    PwmDeInit();   
-}
-
-
-extern void pwm_Set(uint16_t u, uint16_t v, uint16_t w)
-{
-    if(true == _pwm_internals.calibrating)
-    {        
-        u = v = w = 0;
-    }
-    else
-    {   
-        /* Reduce PWM values in counter units */
-        u = PWM_COMPARE_VALUE(u);
-        v = PWM_COMPARE_VALUE(v);
-        w = PWM_COMPARE_VALUE(w);
-    }
-    
-    /* Must be interrupt-safe */
-    _pwm_internals.pwmcomparevalue_protect(true);  
-    PwmComparePhaseU = u;
-    PwmComparePhaseV = v;
-    PwmComparePhaseW = w;
-    _pwm_internals.pwmcomparevalue_protect(false);   
-}
-
-
-
-#endif
+#endif // #if defined(MOTORHAL_changes)
 
 
 /* END OF FILE ********************************************************************************************************/
