+++ emBODY/eBcode/arch-arm/board/amc2c/bsp/motorhal/pwm.h
@@ -55,26 +55,10 @@ extern void PwmDeInit(void);
 
 extern void PwmTest(void);
 
+
 #if defined(MOTORHAL_changes)
-
-
-struct pwm_Configuration
-{
-    enum class START { immediate };
-    
-    START start { START::immediate };
-        
-    constexpr pwm_Configuration() = default;
-    constexpr bool isvalid() const { return true; }
-};
-
-extern void pwm_Init(const pwm_Configuration &config);
-extern void pwm_DeInit();
-
-extern void pwm_Set(uint16_t u, uint16_t v, uint16_t w);
-
-
-#endif
+    // nothing is required
+#endif // #if defined(MOTORHAL_changes) 
 
 #ifdef __cplusplus
     } /* extern "C" */
